Add Pathfinding for ObjectType, AI gets Job Requirements
Other Changes
- Removed Testing code from InputController which was added in commit b57dc0bfd82fbcf87f7e5fb296fa97a568033030
- Removed the FindPathToClosestTile method from commit b57dc0bfd82fbcf87f7e5fb296fa97a568033030
- Add Contains(type, amnt) method to the Inventory class which checks if it has an stack of type and the asked amount
- Job::GetRequirements will return an empty Dictionary is Requirements is null
- LooseItem::RemoveFromStack will makes its Tiles LooseItem null if it is empty and has a Tile
- Changed some Node variables to floats instead of ints

+++ Assets/Scripts/Data/Pathfinding/Node.cs
@@ -7,21 +7,21 @@ public class Node : IHeapItem<Node>{
     public readonly Tile tile;
 
     public readonly bool isWalkable;
-    public readonly int movementPenalty;
+    public readonly float movementPenalty;
 
-    public int gCost, hCost;
+    public float gCost, hCost;
 
     public Node parent;
 
     int heapIndex;
 
-    public Node(Tile tile, bool _isWalkable, int _penalty) {
+    public Node(Tile tile, bool _isWalkable, float _penalty) {
         this.tile = tile;
         this.isWalkable = _isWalkable;
         this.movementPenalty = _penalty;
     }
 
-    public int fCost {
+    public float fCost {
         get {
             return gCost + hCost;
         }
